requirements.txt
appnope==0.1.2
astroid==2.6.6
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
black==21.8b0
certifi==2021.5.30
cftime==1.5.0
click==8.0.1
colorama==0.4.4
couchbase==3.1.3
debugpy==1.3.0
decorator==5.0.9
docutils==0.16
durationpy==0.5
future==0.18.2
iniconfig==1.1.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
lxml==4.6.1
mando==0.6.4
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
netCDF4==1.5.7
numpy==1.19.4
packaging==20.9
pandas==1.1.4
parso==0.8.2
pathspec==0.9.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.3.0
plotly==5.3.1
plotly-express==0.4.1
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
Pygments==2.9.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyproj==3.1.0
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==22.1.0
radon==4.3.2
regex==2021.8.28
scipy==1.7.1
six==1.15.0
snakeviz==2.1.0
statsmodels==0.12.2
tenacity==8.0.1
toml==0.10.2
tomli==1.2.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
wcwidth==0.2.5
wrapt==1.12.1
